+++ eta-company-calendar-frontend/src/app/shared/shared.module.ts
@@ -9,7 +9,7 @@ import { MainLayoutComponent } from '~/app/shared/pages/main-layout.component';
 import { ApiCommunicationService } from '~/app/shared/services/api-communication.service';
 import { ConfigurationService } from '~/app/shared/services/configuration.service';
 import { AuthGuard } from './guards/auth.guard';
-import { AuthInterceptorService } from './services/auth-interceptor.service';
+import { AuthInterceptorService } from './interceptors/auth-interceptor.service';
 import { AuthService } from './services/auth.service';
 
 @NgModule({
